refactor: BGM をルートベースの宣言的単一プレイヤーに変更
mount/unmount で登録・破棄する命令的な方式は、ページ遷移の谷間で一瞬
要求が空になり BGM を破棄→作り直す問題があった（トップ→規約で音が止まる）。

usePathname を真実のソースとし、ルート→曲の対応表(audioManifest)で再生曲を
決める宣言的な単一コントローラに変更。pathname は遷移で連続的に変わるため
谷間が生まれず、同じ曲を割り当てたルート間は Audio を維持して継続再生する。

- AudioController.tsx: layout 常駐の単一プレイヤー（usePathname ベース）
- audioManifest.ts: BGM_BY_ROUTE / resolveBgmKey を追加
- AudioProvider.tsx / useBgm.ts を削除（各ページの BGM 宣言も不要に）

refs #185
