+++ frontend/src/features/diagnosis/components/MbtiSelect.tsx
@@ -6,7 +6,6 @@ import { motion, AnimatePresence } from 'framer-motion';
 import { useSetAtom } from 'jotai';
 import { mbtiAtom, diagnosisStepAtom } from '@/stores/diagnosis';
 import { MBTI_TYPES, MBTI_GROUPS, type MbtiType } from '@/constants/mbti';
-import { useBgm } from '@/components/audio/useBgm';
 
 // タブ・カード・モーダルで共有するジャンル色（分析家 / 外交官 / 番人 / 探検家）
 const GROUP_COLORS = [
@@ -34,9 +33,6 @@ export default function MbtiSelect() {
   const [groupIndex, setGroupIndex] = useState(0);
   const [selected, setSelected] = useState<string | null>(null);
   const [skipModalOpen, setSkipModalOpen] = useState(false);
-
-  // 共通基盤で BGM を再生（トップ・規約ゲームと同じ曲を継続再生する）
-  useBgm('start');
 
   const playSE = useCallback((path: string) => {
     const audio = new Audio(path);
